% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_ordering.R
\name{dist2}
\alias{dist2}
\title{Squared Euclidean distance between two points}
\usage{
dist2(p, q)
}
\arguments{
\item{p, q}{numeric vectors of the same length.}
}
\value{
\code{sum((p - q)^2)}.
}
\description{
Squared Euclidean distance between two points
}
\keyword{internal}
